Refactor deletion modal
