--- app/components/Homepage/Games/Games.module.scss
@@ -1,28 +1,0 @@
-@use 'styles' as *;
-
-.title {
-  @extend %text-style-display-md-semibold;
-
-  color: $color-gray-50;
-}
-
-.titleRow {
-  display: flex;
-  justify-content: space-between;
-}
-
-.removeButton {
-  @extend %text-style-display-xs-semibold;
-
-  background-color: $color-error-500;
-  border: none;
-  border-radius: $border-radius-small;
-  color: $color-primary-25;
-  cursor: pointer;
-  padding: $spacing-8 $spacing-16;
-  transition: all 0.3s;
-}
-
-.removeButton:hover {
-  background-color: $color-error-600;
-}
