--- app/components/Homepage/Games/Modal/Modal.module.scss
@@ -1,69 +1,0 @@
-@use 'styles' as *;
-
-%button {
-  @extend %text-style-text-xl-semibold;
-
-  border: none;
-  border-radius: $border-radius-small;
-  cursor: pointer;
-  padding: $spacing-16;
-  transition: all 0.3s;
-  
-  &:hover {
-    transform: translateY(-3%);
-  }
-
-  &:active {
-    transform: translateY(3%);
-  }
-}
-
-.modal {
-  background: $color-gradient-gray-800-700;
-  border-radius: $border-radius-large;
-  color: $color-gray-300;
-  display: grid;
-  grid-template-rows: auto minmax(0, 1fr);
-  min-height: 0;
-  overflow: hidden;
-  position: relative;
-  width: 30%;
-}
-
-.header {
-  @extend %text-style-display-md-bold;
-
-  background: $color-gradient-primary-900-700;
-  padding: $spacing-32;
-  text-align: center; 
-}
-
-.buttons {
-  @include grid-cols(2);
-
-  column-gap: $spacing-64;
-  justify-content: center;
-  padding: $spacing-64;
-}
-
-.delete {
-  @extend %button;
-
-  background-color: $color-error-600;
-  color: $color-gray-200;
-
-  &:hover {
-    background-color: $color-error-700; 
-  }
-}
-
-.cancel {
-  @extend %button;
-
-  background-color: $color-primary-600;
-  color: $color-gray-200;
-
-  &:hover {
-    background-color: $color-primary-700; 
-  }
-}
